+++ src/chain/tezos/contracts/TezosContractUtils.ts
@@ -1,0 +1,45 @@
+import * as blakejs from 'blakejs';
+
+import { TezosLanguageUtil } from '../TezosLanguageUtil';
+import { TezosNodeReader } from '../TezosNodeReader';
+
+/**
+ *
+ */
+export namespace TezosContractUtils {
+/**
+     * Gets the contract code at the specified address at the head block and compares it to the known hash of the code. This function processes Micheline format contracts.
+     * 
+     * 
+     * @param server Destination Tezos node.
+     * @param address Contract address to query.
+     */
+    export async function verifyDestination(server: string, address: string, expected: string): Promise<boolean> {
+        const contract = await TezosNodeReader.getAccountForBlock(server, 'head', address);
+
+        if (!!!contract.script) { throw new Error(`No code found at ${address}`); }
+
+        const k = Buffer.from(blakejs.blake2s(JSON.stringify(contract.script.code), null, 16)).toString('hex');
+
+        if (k !== expected) { throw new Error(`Contract code hash "${k}" doesn't match expected ${expected}`); }
+
+        return true;
+    }
+
+    /**
+     * In contrast to verifyDestination, this function uses compares Michelson hashes.
+     * 
+     * @param script 
+     */
+    export function verifyScript(script: string, expected: string): boolean {
+        const k = Buffer.from(blakejs.blake2s(TezosLanguageUtil.preProcessMichelsonScript(script).join('\n'), null, 16)).toString('hex');
+
+        if (k !== expected) { throw new Error(`Contract code hash "${k}" doesn't match expected ${expected}`); }
+
+        return true;
+    }
+
+    export function clearRPCOperationGroupHash(hash: string) {
+        return hash.replace(/\"/g, '').replace(/\n/, '');
+    }
+}