+++ src/index-web.ts
@@ -16,7 +16,7 @@ export * from './chain/tezos/contracts/MurbardMultisigHelper';
 export * from './chain/tezos/contracts/StakerDAOTokenHelper';
 export * from './chain/tezos/contracts/TCFBakerRegistryHelper';
 export * from './chain/tezos/contracts/Tzip7ReferenceTokenHelper';
-export * from './chain/tezos/contracts/tzip12/NFTReferenceTokenHelper';
+export * from './chain/tezos/contracts/tzip12/MultiAssetTokenHelper';
 export * from './chain/tezos/contracts/tzip12/SingleAssetTokenHelper';
 export * from './chain/tezos/contracts/TzbtcTokenHelper';